requirements.txt
altgraph==0.17
importlib-metadata==3.4.0
kafka-python==2.0.2
macholib==1.14
modulegraph==0.18
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
typing-extensions==3.7.4.3
zipp==3.4.0
